+++ src/main/java/fr/maxlego08/essentials/commands/commands/spawn/CommandSpawn.java
@@ -4,10 +4,12 @@ import fr.maxlego08.essentials.api.EssentialsPlugin;
 import fr.maxlego08.essentials.api.commands.CommandResultType;
 import fr.maxlego08.essentials.api.commands.Permission;
 import fr.maxlego08.essentials.api.messages.Message;
+import fr.maxlego08.essentials.api.user.User;
 import fr.maxlego08.essentials.module.modules.SpawnModule;
 import fr.maxlego08.essentials.storage.ConfigStorage;
 import fr.maxlego08.essentials.zutils.utils.commands.VCommand;
 import org.bukkit.Location;
+import org.bukkit.entity.Player;
 
 public class CommandSpawn extends VCommand {
 
@@ -16,11 +18,17 @@ public class CommandSpawn extends VCommand {
         this.setModule(SpawnModule.class);
         this.setPermission(Permission.ESSENTIALS_SPAWN);
         this.setDescription(Message.DESCRIPTION_SPAWN);
-        this.onlyPlayers();
+        this.addOptionalArg("player");
     }
 
     @Override
     protected CommandResultType perform(EssentialsPlugin plugin) {
+
+        Player player = this.argAsPlayer(0, this.player);
+        if (player == null) return CommandResultType.SYNTAX_ERROR; // Only if its console
+
+        User user = this.plugin.getUser(player.getUniqueId());
+        if (user == null) return CommandResultType.SYNTAX_ERROR; // Only if its console
 
         Location location = ConfigStorage.spawnLocation;
         if (location == null) {
@@ -28,7 +36,10 @@ public class CommandSpawn extends VCommand {
             return CommandResultType.DEFAULT;
         }
 
-        this.user.teleport(location, Message.TELEPORT_MESSAGE_SPAWN, Message.TELEPORT_SUCCESS_SPAWN);
+        user.teleport(location, Message.TELEPORT_MESSAGE_SPAWN, Message.TELEPORT_SUCCESS_SPAWN);
+        if (this.user == null || user != this.user) {
+            message(sender, Message.TELEPORT_MESSAGE_SPAWN_CONSOLE, player);
+        }
 
         return CommandResultType.SUCCESS;
     }